pyproject.toml
[tool.poetry]
name = "KGB_AFIRM"
version = "0.1.0"
description = "A bot with 'counsellor-like' functions, including callable random affirmations and positive reinforcement."
authors = []
license = "GPL-2.0"

[tool.poetry.dependencies]
python = "~3.7.0"
wheel = "^0.34.2"
setuptools = "^49.2.0"
aiohttp = "^3.6.2"
beautifulsoup4 = "^4.9.1"
pyyaml = "^5.3.1"
aresponses = "^2.0.0"
"discord.py" = "^1.3.4"
boltons = "^20.2.0"
gspread = "^3.6.0"
requests = "^2.24.0"
python-dotenv = "^0.14.0"
praw = "^7.1.0"
oauth2client = "^4.1.3"

[tool.poetry.dev-dependencies]
pytest = "~5.4"
pytest-cov = "^2.10.0"
dpytest = {url= "https://files.pythonhosted.org/packages/57/2d/448a1da35783c1cb13f5168e1d9fd1f719aec1980351adb5b1961f8f6b1c/dpytest-0.0.20-py3-none-any.whl"}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
